pyproject.toml
[project]
name = "superclaw"
version = "0.1.0"
description = "Agent Security Testing Framework - Break AI coding agents with CodeOptiX and Bloom"
authors = [
    { name = "Superagentic AI", email = "team@super-agentic.ai" }
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
keywords = [
    "ai",
    "agent",
    "security",
    "testing",
    "llm",
    "prompt-injection",
    "red-team",
    "openclaw",
    "codeoptix",
    "bloom"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "litellm>=1.40.0",
]

[project.optional-dependencies]
codeoptix = [
    "codeoptix>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]
all = [
    "superclaw[codeoptix,docs]",
]

[project.scripts]
superclaw = "superclaw.cli:app"

[project.urls]
Homepage = "https://github.com/SuperagenticAI/superclaw"
Documentation = "https://superagenticai.github.io/superclaw/"
Repository = "https://github.com/SuperagenticAI/superclaw"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/superclaw"]

# ============================================================
# UV Configuration
# ============================================================

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[tool.uv.sources]
codeoptix = { path = "../superqode/Ignore/codeoptix", editable = true }

# ============================================================
# Ruff Configuration
# ============================================================

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============================================================
# Pytest Configuration
# ============================================================

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

# ============================================================
# MyPy Configuration
# ============================================================

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
